Include dependencies for each section in pages
